%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  QDNAseq.mm10.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{mm10.1kbp.SR50}
\alias{mm10.1kbp.SR50}
\docType{data}


 \title{Mm10 1kbp bin annotations}

 \description{ The annotated data frame contains pre calculated data for
 each bin describing:
   * chromosome: Chromosome name,
   * start: Base pair start position,
   * end: Base pair end position,
   * bases: Percentage of non-N nucleotides (of full bin size),
   * gc: Percentage of C and G nucleotides (of non-N nucleotides),
   * mappability: Average mappability of 50mers with a maximum of 2 mismatches,
   * blacklist: Percent overlap with ENCODE blacklisted regions,
   * residual: Median loess residual from 1000 Genomes (50mers),
   * use: Whether the bin should be used in subsequent analysis steps),
 }

 \author{Daoud Sie}
\keyword{datasets}
